requirements.txt
chainlit
dspy
python-dotenv
